debuged JS and brought some logics

+++ database/migrations/2018_03_20_105126_add_preorder_status_column_to_products_table.php
@@ -14,7 +14,7 @@ class AddPreorderStatusColumnToProductsTable extends Migration
     public function up()
     {
         Schema::table("products", function (Blueprint $table){
-           $table->unsignedInteger('preorder')->nullable();
+           $table->tinyInteger('preorder')->nullable();
         });
     }
